requirements.txt
pre-commit
ruff
mypy
pytest
pytest-cov
pytest-xdist
pymarkdownlnt
pydantic
python-docx
spacy
fastapi
uvicorn
